Move FEPointEvaluation from fe to matrix_free
