+++ tests/matrix_free/point_evaluation_01.cc
@@ -14,24 +14,23 @@
 // ---------------------------------------------------------------------
 
 
-// check FEPointEvaluation for scalar FE_DGQ and MappingFEField by comparing
-// to the output of FEValues with the same settings
+// check FEPointEvaluation for scalar FE_Q and MappingQGeneric by comparing to
+// the output of FEValues with the same settings
 
 #include <deal.II/base/function_lib.h>
 
 #include <deal.II/dofs/dof_handler.h>
 
-#include <deal.II/fe/fe_dgq.h>
-#include <deal.II/fe/fe_point_evaluation.h>
 #include <deal.II/fe/fe_q.h>
-#include <deal.II/fe/fe_system.h>
 #include <deal.II/fe/fe_values.h>
-#include <deal.II/fe/mapping_fe_field.h>
+#include <deal.II/fe/mapping_q_generic.h>
 
 #include <deal.II/grid/grid_generator.h>
 #include <deal.II/grid/tria.h>
 
 #include <deal.II/lac/vector.h>
+
+#include <deal.II/matrix_free/fe_point_evaluation.h>
 
 #include <deal.II/numerics/vector_tools.h>
 
@@ -41,7 +40,7 @@
 
 
 
-template <int dim>
+template <int dim, typename Number = double>
 void
 test(const unsigned int degree)
 {
@@ -53,14 +52,7 @@ test(const unsigned int degree)
   else
     GridGenerator::subdivided_hyper_cube(tria, 2, 0, 1);
 
-  FESystem<dim>   fe_grid(FE_Q<dim>(degree), dim);
-  DoFHandler<dim> dof_handler_grid(tria);
-  dof_handler_grid.distribute_dofs(fe_grid);
-  const ComponentMask mask(dim, true);
-  Vector<double>      location_vector(dof_handler_grid.n_dofs());
-  VectorTools::get_position_vector(dof_handler_grid, location_vector, mask);
-  MappingFEField<dim> mapping(dof_handler_grid, location_vector, mask);
-
+  MappingQGeneric<dim> mapping(degree);
   deallog << "Mapping of degree " << degree << std::endl;
 
   std::vector<Point<dim>> unit_points;
@@ -68,11 +60,11 @@ test(const unsigned int degree)
     {
       Point<dim> p;
       for (unsigned int d = 0; d < dim; ++d)
-        p[d] = static_cast<double>(i) / 17. + 0.015625 * d;
+        p[d] = static_cast<Number>(i) / 17. + 0.015625 * d;
       unit_points.push_back(p);
     }
 
-  FE_DGQ<dim>   fe(degree);
+  FE_Q<dim>     fe(degree);
   FEValues<dim> fe_values(mapping,
                           fe,
                           Quadrature<dim>(unit_points),
@@ -80,22 +72,26 @@ test(const unsigned int degree)
 
   DoFHandler<dim> dof_handler(tria);
   dof_handler.distribute_dofs(fe);
-  Vector<double> vector(dof_handler.n_dofs());
+  Vector<Number> vector(dof_handler.n_dofs());
 
-  FEPointEvaluation<1, dim> evaluator(mapping,
-                                      fe,
-                                      update_values | update_gradients);
+  FEPointEvaluation<1, dim, dim, Number> evaluator(
+    mapping, fe, update_values | update_gradients);
 
-  Tensor<1, dim> exponents;
+  Tensor<1, dim, Number> exponents;
   exponents[0] = 1.;
   VectorTools::interpolate(mapping,
                            dof_handler,
-                           Functions::Monomial<dim>(exponents),
+                           Functions::Monomial<dim, Number>(exponents),
                            vector);
 
-  std::vector<double>         solution_values(fe.dofs_per_cell);
-  std::vector<double>         function_values(unit_points.size());
-  std::vector<Tensor<1, dim>> function_gradients(unit_points.size());
+  std::vector<Number>                 solution_values(fe.dofs_per_cell);
+  std::vector<Number>                 function_values(unit_points.size());
+  std::vector<Tensor<1, dim, Number>> function_gradients(unit_points.size());
+
+  // For float numbers that are sensitive to roundoff in the numdiff
+  // tolerances (absolute 1e-8), we multiply by 1e-3 to ensure that the test
+  // remains robust
+  const double factor_float = std::is_same<Number, float>::value ? 0.001 : 1.;
 
   for (const auto &cell : dof_handler.active_cell_iterators())
     {
@@ -114,9 +110,12 @@ test(const unsigned int degree)
       deallog << "Cell with center " << cell->center(true) << std::endl;
       for (unsigned int i = 0; i < function_values.size(); ++i)
         deallog << mapping.transform_unit_to_real_cell(cell, unit_points[i])
-                << ": " << evaluator.get_value(i) << " error value "
-                << function_values[i] - evaluator.get_value(i) << " error grad "
-                << (evaluator.get_gradient(i) - function_gradients[i]).norm()
+                << ": " << factor_float * evaluator.get_value(i)
+                << " error value "
+                << factor_float * (function_values[i] - evaluator.get_value(i))
+                << " error grad "
+                << factor_float *
+                     (evaluator.get_gradient(i) - function_gradients[i]).norm()
                 << std::endl;
       deallog << std::endl;
 
@@ -130,7 +129,7 @@ test(const unsigned int degree)
                           EvaluationFlags::values | EvaluationFlags::gradients);
 
       for (const auto i : solution_values)
-        deallog << i << " ";
+        deallog << factor_float * i << " ";
       deallog << std::endl;
     }
 }
@@ -144,7 +143,11 @@ main()
   deallog << std::setprecision(10);
 
   test<1>(3);
+  test<2>(1);
   test<2>(2);
   test<2>(6);
+  test<3>(1);
   test<3>(5);
+
+  test<3, float>(5);
 }
