+++ tests/matrix_free/point_evaluation_05.cc
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 2021 by the deal.II authors
+// Copyright (C) 2020 - 2021 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
@@ -14,22 +14,24 @@
 // ---------------------------------------------------------------------
 
 
-// check FEPointEvaluation::real_point(), FEPointEvaluation::unit_point(),
-// FEPointEvaluation::jacobian(), FEPointEvaluation::inverse_jacobian(),
-// FEPointEvaluation::get_unit_gradient().
+// check FEPointEvaluation for scalar FE_Q and MappingField by comparing to
+// the output of FEValues with the same settings
 
 #include <deal.II/base/function_lib.h>
 
 #include <deal.II/dofs/dof_handler.h>
 
-#include <deal.II/fe/fe_point_evaluation.h>
 #include <deal.II/fe/fe_q.h>
-#include <deal.II/fe/mapping_q_generic.h>
+#include <deal.II/fe/fe_system.h>
+#include <deal.II/fe/fe_values.h>
+#include <deal.II/fe/mapping_fe_field.h>
 
 #include <deal.II/grid/grid_generator.h>
 #include <deal.II/grid/tria.h>
 
 #include <deal.II/lac/vector.h>
+
+#include <deal.II/matrix_free/fe_point_evaluation.h>
 
 #include <deal.II/numerics/vector_tools.h>
 
@@ -51,11 +53,18 @@ test(const unsigned int degree)
   else
     GridGenerator::subdivided_hyper_cube(tria, 2, 0, 1);
 
-  MappingQGeneric<dim> mapping(degree);
+  FESystem<dim>   fe_grid(FE_Q<dim>(degree), dim);
+  DoFHandler<dim> dof_handler_grid(tria);
+  dof_handler_grid.distribute_dofs(fe_grid);
+  const ComponentMask mask(dim, true);
+  Vector<double>      location_vector(dof_handler_grid.n_dofs());
+  VectorTools::get_position_vector(dof_handler_grid, location_vector, mask);
+  MappingFEField<dim> mapping(dof_handler_grid, location_vector, mask);
+
   deallog << "Mapping of degree " << degree << std::endl;
 
   std::vector<Point<dim>> unit_points;
-  for (unsigned int i = 0; i < 7; ++i)
+  for (unsigned int i = 0; i < 13; ++i)
     {
       Point<dim> p;
       for (unsigned int d = 0; d < dim; ++d)
@@ -63,11 +72,19 @@ test(const unsigned int degree)
       unit_points.push_back(p);
     }
 
-  FE_Q<dim> fe(degree);
+  FE_Q<dim>     fe(degree);
+  FEValues<dim> fe_values(mapping,
+                          fe,
+                          Quadrature<dim>(unit_points),
+                          update_values | update_gradients);
 
   DoFHandler<dim> dof_handler(tria);
   dof_handler.distribute_dofs(fe);
   Vector<double> vector(dof_handler.n_dofs());
+
+  FEPointEvaluation<1, dim> evaluator(mapping,
+                                      fe,
+                                      update_values | update_gradients);
 
   Tensor<1, dim> exponents;
   exponents[0] = 1.;
@@ -76,15 +93,16 @@ test(const unsigned int degree)
                            Functions::Monomial<dim>(exponents),
                            vector);
 
-  FEPointEvaluation<1, dim> evaluator(mapping,
-                                      fe,
-                                      update_values | update_gradients |
-                                        update_quadrature_points |
-                                        update_jacobians);
-  std::vector<double>       solution_values(fe.dofs_per_cell);
+  std::vector<double>         solution_values(fe.dofs_per_cell);
+  std::vector<double>         function_values(unit_points.size());
+  std::vector<Tensor<1, dim>> function_gradients(unit_points.size());
 
   for (const auto &cell : dof_handler.active_cell_iterators())
     {
+      fe_values.reinit(cell);
+      fe_values.get_function_values(vector, function_values);
+      fe_values.get_function_gradients(vector, function_gradients);
+
       cell->get_dof_values(vector,
                            solution_values.begin(),
                            solution_values.end());
@@ -94,16 +112,26 @@ test(const unsigned int degree)
                          EvaluationFlags::values | EvaluationFlags::gradients);
 
       deallog << "Cell with center " << cell->center(true) << std::endl;
-      for (unsigned int i = 0; i < unit_points.size(); ++i)
-        deallog << "unit point " << unit_points[i] << std::endl
-                << "unit point via evaluator: " << evaluator.unit_point(i)
-                << std::endl
-                << "real point: " << evaluator.real_point(i) << std::endl
-                << "jacobian: " << Tensor<2, dim>(evaluator.jacobian(i))
-                << std::endl
-                << "inverse jacobian: "
-                << Tensor<2, dim>(evaluator.inverse_jacobian(i)) << std::endl
+      for (unsigned int i = 0; i < function_values.size(); ++i)
+        deallog << mapping.transform_unit_to_real_cell(cell, unit_points[i])
+                << ": " << evaluator.get_value(i) << " error value "
+                << function_values[i] - evaluator.get_value(i) << " error grad "
+                << (evaluator.get_gradient(i) - function_gradients[i]).norm()
                 << std::endl;
+      deallog << std::endl;
+
+      for (unsigned int i = 0; i < unit_points.size(); ++i)
+        {
+          evaluator.submit_value(evaluator.get_value(i), i);
+          evaluator.submit_gradient(evaluator.get_gradient(i), i);
+        }
+
+      evaluator.integrate(solution_values,
+                          EvaluationFlags::values | EvaluationFlags::gradients);
+
+      for (const auto i : solution_values)
+        deallog << i << " ";
+      deallog << std::endl;
     }
 }
