+++ tests/matrix_free/point_evaluation_10.cc
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 2020 - 2021 by the deal.II authors
+// Copyright (C) 2021 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
@@ -14,16 +14,16 @@
 // ---------------------------------------------------------------------
 
 
-// check FEPointEvaluation for vector-valued FE_Q and MappingQGeneric by
-// comparing to the output of FEValues with the same settings
+// check FEPointEvaluation for scalar FE_DGQArbitraryNodes and MappingQGeneric
+// by comparing to the output of FEValues with the same settings (apart from
+// the finite element, this test is the same as point_evaluation_02)
 
 #include <deal.II/base/function_lib.h>
+#include <deal.II/base/quadrature_lib.h>
 
 #include <deal.II/dofs/dof_handler.h>
 
-#include <deal.II/fe/fe_point_evaluation.h>
-#include <deal.II/fe/fe_q.h>
-#include <deal.II/fe/fe_system.h>
+#include <deal.II/fe/fe_dgq.h>
 #include <deal.II/fe/fe_values.h>
 #include <deal.II/fe/mapping_q_generic.h>
 
@@ -32,28 +32,13 @@
 
 #include <deal.II/lac/vector.h>
 
+#include <deal.II/matrix_free/fe_point_evaluation.h>
+
 #include <deal.II/numerics/vector_tools.h>
 
 #include <iostream>
 
 #include "../tests.h"
-
-
-
-template <int dim>
-class MyFunction : public Function<dim>
-{
-public:
-  MyFunction()
-    : Function<dim>(dim)
-  {}
-
-  double
-  value(const Point<dim> &p, const unsigned int component) const override
-  {
-    return component + p[component];
-  }
-};
 
 
 
@@ -81,8 +66,8 @@ test(const unsigned int degree)
       unit_points.push_back(p);
     }
 
-  FESystem<dim> fe(FE_Q<dim>(degree), dim);
-  FEValues<dim> fe_values(mapping,
+  FE_DGQArbitraryNodes<dim> fe(QGauss<1>(degree + 1));
+  FEValues<dim>             fe_values(mapping,
                           fe,
                           Quadrature<dim>(unit_points),
                           update_values | update_gradients);
@@ -91,23 +76,26 @@ test(const unsigned int degree)
   dof_handler.distribute_dofs(fe);
   Vector<double> vector(dof_handler.n_dofs());
 
-  FEPointEvaluation<dim, dim> evaluator(mapping,
-                                        fe,
-                                        update_values | update_gradients);
+  FEPointEvaluation<1, dim> evaluator(mapping,
+                                      fe,
+                                      update_values | update_gradients);
 
-  VectorTools::interpolate(mapping, dof_handler, MyFunction<dim>(), vector);
+  Tensor<1, dim> exponents;
+  exponents[0] = 1.;
+  VectorTools::interpolate(mapping,
+                           dof_handler,
+                           Functions::Monomial<dim>(exponents),
+                           vector);
 
   std::vector<double>         solution_values(fe.dofs_per_cell);
-  std::vector<Tensor<1, dim>> function_values(unit_points.size());
-  std::vector<Tensor<2, dim>> function_gradients(unit_points.size());
-
-  FEValuesExtractors::Vector extractor(0);
+  std::vector<double>         function_values(unit_points.size());
+  std::vector<Tensor<1, dim>> function_gradients(unit_points.size());
 
   for (const auto &cell : dof_handler.active_cell_iterators())
     {
       fe_values.reinit(cell);
-      fe_values[extractor].get_function_values(vector, function_values);
-      fe_values[extractor].get_function_gradients(vector, function_gradients);
+      fe_values.get_function_values(vector, function_values);
+      fe_values.get_function_gradients(vector, function_gradients);
 
       cell->get_dof_values(vector,
                            solution_values.begin(),
@@ -121,8 +109,7 @@ test(const unsigned int degree)
       for (unsigned int i = 0; i < function_values.size(); ++i)
         deallog << mapping.transform_unit_to_real_cell(cell, unit_points[i])
                 << ": " << evaluator.get_value(i) << " error value "
-                << (function_values[i] - evaluator.get_value(i)).norm()
-                << " error grad "
+                << function_values[i] - evaluator.get_value(i) << " error grad "
                 << (evaluator.get_gradient(i) - function_gradients[i]).norm()
                 << std::endl;
       deallog << std::endl;
@@ -150,6 +137,8 @@ main()
   initlog();
   deallog << std::setprecision(10);
 
+  test<1>(3);
+  test<2>(1);
   test<2>(2);
   test<2>(6);
   test<3>(5);
