+++ tests/matrix_free/point_evaluation_11.cc
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 2020 - 2021 by the deal.II authors
+// Copyright (C) 2021 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
@@ -14,15 +14,16 @@
 // ---------------------------------------------------------------------
 
 
-// check FEPointEvaluation for scalar FE_Q and MappingQGeneric by comparing to
-// the output of FEValues with the same settings
+// check FEPointEvaluation for scalar FE_DGQHermite and MappingQGeneric by
+// comparing to the output of FEValues with the same settings (apart from the
+// finite element and the interpolated function, this is the same as
+// point_evaluation_02)
 
 #include <deal.II/base/function_lib.h>
 
 #include <deal.II/dofs/dof_handler.h>
 
-#include <deal.II/fe/fe_point_evaluation.h>
-#include <deal.II/fe/fe_q.h>
+#include <deal.II/fe/fe_dgq.h>
 #include <deal.II/fe/fe_values.h>
 #include <deal.II/fe/mapping_q_generic.h>
 
@@ -30,6 +31,8 @@
 #include <deal.II/grid/tria.h>
 
 #include <deal.II/lac/vector.h>
+
+#include <deal.II/matrix_free/fe_point_evaluation.h>
 
 #include <deal.II/numerics/vector_tools.h>
 
@@ -39,7 +42,7 @@
 
 
 
-template <int dim, typename Number = double>
+template <int dim>
 void
 test(const unsigned int degree)
 {
@@ -59,38 +62,32 @@ test(const unsigned int degree)
     {
       Point<dim> p;
       for (unsigned int d = 0; d < dim; ++d)
-        p[d] = static_cast<Number>(i) / 17. + 0.015625 * d;
+        p[d] = static_cast<double>(i) / 17. + 0.015625 * d;
       unit_points.push_back(p);
     }
 
-  FE_Q<dim>     fe(degree);
-  FEValues<dim> fe_values(mapping,
+  FE_DGQHermite<dim> fe(degree);
+  FEValues<dim>      fe_values(mapping,
                           fe,
                           Quadrature<dim>(unit_points),
                           update_values | update_gradients);
 
   DoFHandler<dim> dof_handler(tria);
   dof_handler.distribute_dofs(fe);
-  Vector<Number> vector(dof_handler.n_dofs());
+  Vector<double> vector(dof_handler.n_dofs());
 
-  FEPointEvaluation<1, dim, dim, Number> evaluator(
-    mapping, fe, update_values | update_gradients);
+  FEPointEvaluation<1, dim> evaluator(mapping,
+                                      fe,
+                                      update_values | update_gradients);
 
-  Tensor<1, dim, Number> exponents;
-  exponents[0] = 1.;
-  VectorTools::interpolate(mapping,
-                           dof_handler,
-                           Functions::Monomial<dim, Number>(exponents),
-                           vector);
+  // FE_DGQHermite is not interpolatory, so we just set up some arbitrary
+  // content in the vector for the sake of comparison
+  for (unsigned int i = 0; i < vector.size(); ++i)
+    vector(i) = (i + 2) % 17;
 
-  std::vector<Number>                 solution_values(fe.dofs_per_cell);
-  std::vector<Number>                 function_values(unit_points.size());
-  std::vector<Tensor<1, dim, Number>> function_gradients(unit_points.size());
-
-  // For float numbers that are sensitive to roundoff in the numdiff
-  // tolerances (absolute 1e-8), we multiply by 1e-3 to ensure that the test
-  // remains robust
-  const double factor_float = std::is_same<Number, float>::value ? 0.001 : 1.;
+  std::vector<double>         solution_values(fe.dofs_per_cell);
+  std::vector<double>         function_values(unit_points.size());
+  std::vector<Tensor<1, dim>> function_gradients(unit_points.size());
 
   for (const auto &cell : dof_handler.active_cell_iterators())
     {
@@ -109,12 +106,9 @@ test(const unsigned int degree)
       deallog << "Cell with center " << cell->center(true) << std::endl;
       for (unsigned int i = 0; i < function_values.size(); ++i)
         deallog << mapping.transform_unit_to_real_cell(cell, unit_points[i])
-                << ": " << factor_float * evaluator.get_value(i)
-                << " error value "
-                << factor_float * (function_values[i] - evaluator.get_value(i))
-                << " error grad "
-                << factor_float *
-                     (evaluator.get_gradient(i) - function_gradients[i]).norm()
+                << ": " << evaluator.get_value(i) << " error value "
+                << function_values[i] - evaluator.get_value(i) << " error grad "
+                << (evaluator.get_gradient(i) - function_gradients[i]).norm()
                 << std::endl;
       deallog << std::endl;
 
@@ -128,7 +122,7 @@ test(const unsigned int degree)
                           EvaluationFlags::values | EvaluationFlags::gradients);
 
       for (const auto i : solution_values)
-        deallog << factor_float * i << " ";
+        deallog << i << " ";
       deallog << std::endl;
     }
 }
@@ -142,11 +136,7 @@ main()
   deallog << std::setprecision(10);
 
   test<1>(3);
-  test<2>(1);
   test<2>(2);
   test<2>(6);
-  test<3>(1);
   test<3>(5);
-
-  test<3, float>(5);
 }
